fix some ts build issues
