+++ src/wtsdk/src/cj4/fmc/CJ4_FMC_NavigationService.ts
