+++ src/wtsdk/src/cj4/pfd/CJ4_PFD_BotLeftMessageController.ts
@@ -1,11 +1,11 @@
-import { MessageLevel } from "../../../messages/Message";
-import { MessageController } from "../../../messages/MessageController";
+import { MessageLevel } from "../../messages/Message";
+import { MessageController } from "../../messages/MessageController";
 import { CJ4_PFD_Message } from "./CJ4_PFD_Message";
 
 /**
- * The message controller for the right PFD bottom line
+ * The message controller for the left PFD bottom line
  */
-export class CJ4_PFD_BotRightMessageController extends MessageController<CJ4_PFD_Message> {
+export class CJ4_PFD_BotLeftMessageController extends MessageController<CJ4_PFD_Message> {
 
   constructor() {
     super(CJ4_PFD_Message);
@@ -20,6 +20,13 @@ export class CJ4_PFD_BotRightMessageController extends MessageController<CJ4_PFD
     // get first message
     this._currentMsg = this._messages.values().next().value;
     return `<span class="${(this._currentMsg.level === MessageLevel.Yellow ? "yellow" : "white")} ${(this._currentMsg.shouldBlink() === true) ? "blinking" : ""}">${this._currentMsg.content}</span>`;
+  }
+
+  public getMsgObj(): CJ4_PFD_Message {
+    if (!this.hasMsg()) {
+      return undefined;
+    }
+    return this._messages.values().next().value;
   }
 
   public post(content: string, level: MessageLevel, checkHandler: () => boolean = () => false, blinkCheckHandler: () => boolean = () => false): CJ4_PFD_Message {
