+++ src/wtsdk/src/cj4/pfd/CJ4_PFD_Message.ts
@@ -1,4 +1,4 @@
-import { Message } from "../../../messages/Message";
+import { Message } from "../../messages/Message";
 
 export class CJ4_PFD_Message extends Message {
   private _blinkCheckHandler: () => boolean;
