+++ src/wtsdk/src/cj4/pfd/CJ4_PFD_TopMessageController.ts
@@ -1,4 +1,4 @@
-import { MessageController } from "../../../messages/MessageController";
+import { MessageController } from "../../messages/MessageController";
 import { CJ4_PFD_Message } from "./CJ4_PFD_Message";
 
 /**
