+++ src/wtsdk/src/wtsdk.ts
@@ -1,3 +1,7 @@
-export * from "./flightplanning/FlightPlanManager";
-export * from "./cj4/ui/fmc/pages/CJ4_FMC_MsgPage";
 export * from "./messages/MessageController";
+export * from "./messages/Message";
+export * from "./flightplanning/FlightPlanManager";
+export * from "./cj4/fmc/pages/CJ4_FMC_MsgPage";
+export * from "./cj4/fmc/CJ4_FMC_NavigationService";
+export * from "./cj4/fmc/CJ4_FMC_MessageController";
+export * from "./flightplanning/HoldDetails";
